Added Unittests for Pages, Added h2 in-memory for Unittests

+++ src/main/java/de/eposcat/master/approachImpl/JSON_Oracle_DatabaseAdapter.java
@@ -74,9 +74,9 @@ public Page createPageWithAttributes(String typename, Map<String, Attribute> att
 
 
     @Override
-    public Page loadPage(int pageId) throws SQLException{
+    public Page loadPage(long pageId) throws SQLException{
         PreparedStatement stLoadPage = conn.prepareStatement("SELECT * FROM pages WHERE id = ?");
-        stLoadPage.setInt(1, pageId);
+        stLoadPage.setLong(1, pageId);
         
         ResultSet rsLoadPage = stLoadPage.executeQuery();
